feat(environments): #797 list my environments
