--- web/src/app/environments/list-private/environments-list-private.component.html
@@ -1,3 +1,0 @@
-<div class="mat-content" layout-padding flex>
-  Environment Private List
-</div>
