--- web/src/app/environments/list-private/environments-list-private.component.ts
@@ -1,17 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { EnvironmentService } from '../environment.service';
-
-@Component({
-  selector: 'qs-environments-list-private',
-  templateUrl: './environments-list-private.component.html',
-})
-export class EnvironmentsListPrivateComponent implements OnInit {
-
-  constructor(private environmentService: EnvironmentService) {
-
-  }
-
-  ngOnInit() {
-    this.environmentService.findAllByOwner().subscribe((environments) => console.log(environments));
-  }
-}
