+++ web/src/app/environments/list/environments-list.component.html
@@ -1,3 +1,111 @@
+<div layout-gt-sm="row" tdMediaToggle="gt-xs" [mediaClasses]="['push-sm']">
+  <div layout-gt-xs="row" flex>
+    <div flex-gt-xs="20">
+      <mat-card>
+        <mat-card-content class="text-center">
+          <div class="mat-headline push-bottom-none">{{ summary.environments | digits }}
+            <!--<mat-icon class="tc-green-500 text-sm">arrow_upward</mat-icon>-->
+          </div>
+          <mat-icon class="icon tc-cyan-900">developer_board</mat-icon>
+          <div>
+            <div class="mat-subheading-2 push-bottom-none">Environments</div>
+            <div class="mat-caption tc-grey-600">Public</div>
+          </div>
+        </mat-card-content>
+      </mat-card>
+    </div>
+    <div flex-gt-xs="20">
+      <mat-card>
+        <mat-card-content class="text-center">
+          <div class="mat-headline push-bottom-none">{{ summary.releases | digits }}
+            <!--<mat-icon class="tc-green-500 text-sm">arrow_upward</mat-icon>-->
+          </div>
+          <mat-icon class="icon tc-cyan-900">new_releases</mat-icon>
+          <div>
+            <div class="mat-subheading-2 push-bottom-none">Releases</div>
+            <div class="mat-caption tc-grey-600">Public</div>
+          </div>
+        </mat-card-content>
+      </mat-card>
+    </div>
+    <div flex-gt-xs="20">
+      <mat-card>
+        <mat-card-content class="text-center">
+          <div class="mat-headline push-bottom-none">{{ summary.monitors | digits }}
+            <!--<mat-icon class="tc-green-500 text-sm">arrow_upward</mat-icon>-->
+          </div>
+          <mat-icon class="icon tc-cyan-900">timelapse</mat-icon>
+          <div>
+            <div class="mat-subheading-2 push-bottom-none">Monitors</div>
+            <div class="mat-caption tc-grey-600">Public</div>
+          </div>
+        </mat-card-content>
+      </mat-card>
+    </div>
+    <div flex-gt-xs="20">
+      <mat-card>
+        <mat-card-content class="text-center">
+          <div class="mat-headline push-bottom-none">{{ summary.pings | digits }}
+            <!--<mat-icon class="tc-green-500 text-sm">arrow_upward</mat-icon>-->
+          </div>
+          <mat-icon class="icon tc-cyan-900">receipt</mat-icon>
+          <div>
+            <div class="mat-subheading-2 push-bottom-none">Pings</div>
+            <div class="mat-caption tc-grey-600">Public</div>
+          </div>
+        </mat-card-content>
+      </mat-card>
+    </div>
+    <div flex-gt-xs="20">
+      <mat-card>
+        <mat-card-content class="text-center">
+          <div class="mat-headline push-bottom-none">{{ summary.views | digits }}
+            <!--<mat-icon class="tc-green-500 text-sm">arrow_upward</mat-icon>-->
+          </div>
+          <mat-icon class="icon tc-cyan-900">record_voice_over</mat-icon>
+          <div>
+            <div class="mat-subheading-2 push-bottom-none">Views</div>
+            <div class="mat-caption tc-grey-600">Public</div>
+          </div>
+        </mat-card-content>
+      </mat-card>
+    </div>
+  </div>
+</div>
+
 <div class="mat-content" layout-padding flex>
-  Environment List
+  <mat-card>
+    <mat-card-title>Public Environments</mat-card-title>
+    <mat-card-subtitle>Recently active</mat-card-subtitle>
+    <mat-divider></mat-divider>
+    <mat-nav-list class="will-load item-list">
+      <ng-template tdLoading="items.load">
+        <ng-template let-environment let-last="last" ngFor [ngForOf]="environments.list">
+          <a mat-list-item layout-align="row" [routerLink]="['/environments', environment.id, 'view']">
+            <mat-icon matListAvatar *ngIf="!environment.logo">developer_board</mat-icon>
+            <img matListAvatar [src]="environment.logo" *ngIf="environment.logo" [alt]="environment.title">
+            <h3 matLine>{{ environment.title }}</h3>
+            <p matLine>{{ environment.description | truncate:50 }}</p>
+            <span flex></span>
+            <span class="mat-caption text-right" flex="20" layout-align="row">
+              <p>{{ environment.type }}</p>
+            </span>
+            <span class="mat-caption text-right" flex="20" layout-align="row" *ngIf="environment.latestRelease.release">
+              <h3>{{ environment.latestRelease.release | digits }}</h3>
+              <p>{{ environment.latestRelease.updatedAt | timeAgo }}</p>
+            </span>
+            <span class="mat-caption text-right" flex="20" layout-align="row" *ngIf="!environment.latestRelease.release">
+              <p>No releases</p>
+            </span>
+            <span class="mat-caption text-right" flex="20" layout-align="row">
+              <p>{{ environment.releases }} releases</p>
+              <p>{{ environment.pings.valid + environment.pings.invalid | digits }} pings</p>
+            </span>
+          </a>
+          <mat-progress-bar mode="determinate" [value]="environment.progress.current"></mat-progress-bar>
+          <mat-divider *ngIf="!last" [inset]="true"></mat-divider>
+        </ng-template>
+      </ng-template>
+    </mat-nav-list>
+  </mat-card>
 </div>
