+++ web/src/app/environments/list/public/environments-list-public.component.html
@@ -1,0 +1,8 @@
+<mat-toolbar>
+  <mat-toolbar-row>
+    <span flex></span>
+    <button mat-button (click)="refresh()"><mat-icon>refresh</mat-icon></button>
+  </mat-toolbar-row>
+</mat-toolbar>
+
+<qs-environments-list [environments]="environments"></qs-environments-list>
