+++ web/src/app/environments/list/public/public.environments.resolver.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
 import { Resolve } from '@angular/router';
-import { EnvironmentService } from './environment.service';
 import { Observable } from 'rxjs/Observable';
-import { List } from '../list';
-import { Environment } from './environment.model';
+import { List } from "../../../list";
+import { Environment } from "../../environment.model";
+import { EnvironmentService } from "../../environment.service";
 
 @Injectable()
 export class PublicEnvironmentsResolver implements Resolve<List<Environment>> {
